creating plotting utils module for plotting functions
